lambda-ecs/api/download-news/requirements.txt
requests==2.31.0
newspaper3k==0.2.8
lxml[html_clean]==5.2.1